Huge update. Rewrote entire readme, building and knowissues. Edited the python code to be a little more structured. Added an Innosetup file to create and easy installer. (No more manual copying)
