+++ knownissues.md
@@ -1,37 +1,27 @@
-<!-- Error:
-```sh
-error here
-```
-<details>
-  <summary>Solution</summary>
-  <p> If there is any </p>
-  <p> Reason: </p>
-</details>
+### Known Issues and Solutions
 
---- -->
-### 1. Error:
-Error:
+Here is a list of known issues and their common solutions:
+
+1. **Error:**
+
 ```sh
 ValueError: y1 must be greater than or equal to y0
 ```
+
 <details>
   <summary>Solution</summary>
-  <p> Check that you are using the latest script, last update 14.08.2023 9:45 PM GMT+2 </p>
-  <p> Reason: The "pillow" package version I used was slightly old, therefore I needed to fix the code for the new 10.0.0 version of "pillow".</p>
+  Ensure you are using the latest script (last update: 14.08.2023, 9:45 PM GMT+2). The issue arises because the "pillow" package version used was slightly old. The code has been fixed for the new 10.0.0 version of "pillow".
 </details>
 
 ---
 
-2. Error:
+2. **Error:**
+
 ```sh
 Battery always N/A?
 ```
+
 <details>
   <summary>Solution</summary>
-  <p> Reconnect the dongle to your pc, pull it out and put it together and then connect again to your PC via USB. </p>
-  <p> Reason: Connection might have lost info over time idk xD Caching errors</p>
+  Reconnect the dongle to your PC. Remove it and then reconnect it to your PC via USB. The issue might be due to connection information being lost over time or caching errors.
 </details>
-
-
-
-
