+++ readme.md
@@ -1,82 +1,91 @@
-## Steelseries Mouse Battery retrieval
+# Steelseries Mouse Battery Retrieval
 
-This is a cool Windows app to retrieve the battery level of your Steelseries
-mouse and display it in the system tray.
+This is a handy Windows application that retrieves the battery level of your Steelseries mouse and displays it in the system tray.
+
+## Table of Contents
+
+- [Usage](#usage)
+- [Tested Devices](#tested-devices)
+- [Installation](#installation)
+  - [Latest Version (Recommended)](#latest-version-recommended)
+- [Building from Source](#building-application-and-installer-from-source)
+- [Supported Devices](#supported-devices)
+- [Troubleshooting](#troubleshooting)
+- [Uninstallation](#uninstallation)
+- [Acknowledgements](#acknowledgements)
+- [License](#license)
 
 ## Usage
 
-Once started, you can hover over the icon to see the battery level. On right
-click, you can exit the app or see the battery level again.
+Once the application is started, you can hover over the icon to see the battery level. Right-clicking allows you to exit the application or view the battery level again.
 
-## Tested on
+## Tested Devices
 
 - Steelseries AEROX 3 Wireless (2.4G mode)
-- Steelseries Prime Wireless
 - Steelseries AEROX 9 Wireless (2.4G mode)
+- Steelseries Prime Wireless
 
-## Installation 
-0. Download the newest App from the [releases](https://github.com/yurtemre7/mouse-battery/releases/) tab.
-1. Put the `mouse.exe` file in your startup folder 
-   (C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Startup).
-2. Start the program manually or restart your PC.
+## Installation
 
-## Installation (old)
-0. Download the git repository as a zip file and extract it somewhere (or clone
-   it with git)
-1. Install Python 3 (https://www.python.org/downloads/) and make sure to check
-   the box to add it to your PATH.
-2. Pip install the following packages:
-   - `pip install rivalcfg==4.11.0`
-   - `pip install pystray==0.19.5`
-   - `pip install pillow==10.0.0`
-3. Run the script
-   - Put the `start_mouse.bat` file as a shortcut in your startup folder
-     (C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Startup).
-   - You may need to change the bat script to point to your python installation
-     if you have multiple versions installed or renamed the executable.
-   - or you can run the script manually: `python3 mouse.py`
+### Latest Version (Recommended)
 
-## Supported by the `rivalcfg` library
+1. Download the latest application installer from the [Releases](https://github.com/yurtemre7/mouse-battery/releases/) tab.
+2. Run the installer. This will install the application and place a shortcut in your Start Menu and add it to your auto-startup folder.
+3. After installation, the installer will ask you if you want to run the application. If you choose not to, you can run it from the Start Menu shortcut or by restarting your computer.
 
-- [rivalcfg supported devices](https://flozz.github.io/rivalcfg/devices/index.html)
+#### Manual installation
 
-### Problems?
+1. Download the Git repository as a zip file and extract it somewhere (or clone it with Git).
+2. Install Python 3 (<https://www.python.org/downloads/>) and ensure you check the box to add it to your PATH.
+3. Install the following packages with Pip:
+   - `pip install rivalcfg`
+   - `pip install pystray`
+   - `pip install pillow`
+4. Run the script:
+   - Place the `start_mouse.bat` file as a shortcut in your startup folder (C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Startup).
+   - You may need to modify the bat script to point to your Python installation if you have multiple versions installed or have renamed the executable.
+   - Alternatively, you can run the script manually.
 
-Firstly look at the `knownissues.md` file in the folder to see if your problem is listed.
+## Building application and installer from source
 
-Run the script in the `mouse-battery` folder:
+To build the application yourself, check out [`building.md`](https://github.com/yurtemre7/mouse-battery/blob/master/building.md)
+
+## Supported Devices
+
+The `rivalcfg` library supports a variety of devices. A complete list of supported devices can be found [here](https://flozz.github.io/rivalcfg/devices/index.html).
+
+## Troubleshooting
+
+If you encounter any issues, first check the [`knownissues.md`](https://github.com/yurtemre7/mouse-battery/blob/master/knownissues.md) file in the repository to see if your problem is already listed.
+
+If your problem is not listed, you can run the script in the `mouse-battery` folder. Note the output of the script and open an issue with the output and your mouse model. Provide additional details, such as whether you have multiple mice connected, your connection mode, etc.
 
 ```sh
-python3 mouse.py
+python Steel_battery_status.py
 ```
 
-Get the output of the script and open an issue with the output and your mouse
-model. Give extra detail, e.g. if you have multiple mice connected, your
-connection mode etc.
+## Uninstallation
 
-## Uninstall guide
-Sorry to see you go :(
-1. Delete the `mouse.exe` file in your startup folder
-   (C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Startup)
-2. Yeah that's it, lol.
+### Latest Version
 
-## Uninstall guide (old)
-Sorry to see you go :(
-1. Delete the `mouse-battery` folder
-2. Delete the `start_mouse.bat` file in your startup folder
-   (C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Startup)
-3. Delete the python libraries installed via pip:
+1. Go to the Control Panel and select "Uninstall a program".
+2. Find "Steel_battery_status" in the list and click "Uninstall".
+
+### Older Version
+
+1. Delete the `mouse-battery` folder.
+2. Delete the `start_mouse.bat` file in your startup folder (C:\ProgramData\Microsoft\Windows\Start Menu\Programs\Startup).
+3. Delete the Python libraries installed via Pip:
    - `pip uninstall rivalcfg`
    - `pip uninstall pystray`
    - `pip uninstall pillow`
-4. Uninstall Python 3 if you don't need it anymore
+4. Uninstall Python 3 if you no longer need it.
 
-### Thanks to
+## Acknowledgements
 
-- [DeveloperX19](https://github.com/DeveloperX19) for the license of his
-  intellectual property of the art of the icon.
-- [flozz](https://github.com/flozz) for the `rivalcfg` library and the idea of a standalone python executable. 
+- [DeveloperX19](https://github.com/DeveloperX19) for the license of his intellectual property of the icon art.
+- [flozz](https://github.com/flozz) for the `rivalcfg` library and the idea of a standalone Python executable.
 
 ## License
 
-MIT: Feel free to use this code as you wish.
+MIT: Feel free to use this code as you wish. If you do use it, I'd appreciate a mention.
